requirements.txt
fastapi
uvicorn
transformers
torch
pytesseract
Pillow
streamlit
requests
python-multipart
